Refactored ConfigManager under Model package
