+++ src/main/java/seedu/duke/model/ModelMain.java
@@ -1,12 +1,15 @@
 package seedu.duke.model;
 
 //@@author AndreWongZH
+
+import seedu.duke.model.config.ConfigManager;
+
 /**
  * Represents the main model class which is inherited by all Managers.
  *
  * @see seedu.duke.model.event.EventManager
  * @see ModelManager
- * @see seedu.duke.ui.ConfigManager
+ * @see ConfigManager
  */
 public abstract class ModelMain {
 }
